% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDynamo.R
\name{Dynamo}
\alias{Dynamo}
\title{Dynamo Function}
\usage{
Dynamo(testtable, numEnc, numTask, userNumEncodings)
}
\arguments{
\item{testtable}{An input table with dimension-tasks as columns and visual encodings as rows, with each corresponding cell containing a rank.}

\item{numEnc}{The number of encodings}

\item{numTask}{The number of tasks}

\item{userNumEncodings}{The number of encoding assignments required to be assigned}
}
\description{
This function accepts an input table, and a number of other parameters that define the bounds of the linear program run as part of Dynamo.
}
\examples{
Dynamo()
}
\keyword{Dynamo,}
\keyword{linear}
\keyword{programming}

