% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rDynamo.R
\name{linprog}
\alias{linprog}
\title{Linprog Function}
\usage{
linprog(inputMatrix)
}
\arguments{
\item{inputMatrix}{An input table with dimension-tasks as columns and visual encodings as rows, with each corresponding cell containing a rank.}
}
\description{
This linear program function accepts an input table object as a parameter. The function solves the linear program. The constraints cannot be parameterized in the function via arguments.
}
\examples{
linprog()
}
\keyword{Dynamo,}
\keyword{linear}
\keyword{programming}

